feat: 考勤管理系统 v1.0
- 三角色系统：前台(user) + 后台(admin) + 工作台(workstation)
- GPS打卡 + 高德地图逆地理编码（坐标转地名）
- 工作台代打卡 + 实时活动流监控
- 请假/加班/补卡申请 + 审批流程
- 考勤规则/部门/用户/公司信息管理
- 数据看板（统计面板）
- 移动端响应式适配
- 操作日志（文件+数据库）
- 通知提醒（申请状态变更）
- Jackson JSR310 时间序列化修复
- 全局异常处理（HTTP状态码规范化）
- 简约现代UI设计

+++ src/main/java/com/attendance/entity/Attendance.java
@@ -3,8 +3,8 @@ package com.attendance.entity;
 import java.time.LocalDateTime;
 
 public class Attendance {
-    private Long id;
-    private Long userId;
+    private Integer id;
+    private Integer userId;
     private Integer type;          // 1=上班打卡, 2=下班打卡
     private Double latitude;       // 打卡纬度
     private Double longitude;      // 打卡经度
@@ -13,10 +13,10 @@ public class Attendance {
     private String remark;         // 备注（异常原因等）
     private LocalDateTime createTime;
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-    public Long getUserId() { return userId; }
-    public void setUserId(Long userId) { this.userId = userId; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
+    public Integer getUserId() { return userId; }
+    public void setUserId(Integer userId) { this.userId = userId; }
     public Integer getType() { return type; }
     public void setType(Integer type) { this.type = type; }
     public Double getLatitude() { return latitude; }
